pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lexgraph"
version = "0.1.0"
description = "Graph-first, citation-first Indian legal research system"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Abhishek" }]
dependencies = [
    # core
    "fastapi>=0.115",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "typer>=0.12",

    # graph / db
    "graphiti-core>=0.3",
    "neo4j>=5.23",
    "redis>=5.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",

    # storage / queue
    "minio>=7.2",

    # parsing / nlp
    "pypdf>=4.3",
    "pdfminer.six>=20240706",
    "python-docx>=1.1",
    "lxml>=5.2",
    "beautifulsoup4>=4.12",
    "regex>=2024.5.15",
    "rapidfuzz>=3.9",
    "python-magic>=0.4",
    "tqdm>=4.66",
    "pyyaml>=6.0",

    # ml (optional at runtime; imported lazily)
    "sentence-transformers>=3.0",
    "transformers>=4.44",
    "torch>=2.2",
    "qdrant-client>=1.11",
    "openai>=1.40",
    # hf datasets / hub (used by scripts/fetch_hf_corpora.py to seed
    # public-adapter caches from HuggingFace-hosted legal corpora)
    "datasets>=2.20",
    "huggingface-hub>=0.24",

    # ocr (optional at runtime)
    "pytesseract>=0.3.10",
    "pdf2image>=1.17",
]

[project.optional-dependencies]
# Codebase knowledge graph: PyPI ``graphifyy``, CLI ``graphify`` (see scripts/build_project_graph.sh)
graphify = ["graphifyy>=0.4"]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "types-requests",
    "httpx",
    "respx>=0.21",
]

[project.scripts]
lexgraph = "services.cli:app"
lexgraph-api = "services.svc_http.main:run"
lexgraph-worker-ingest = "services.svc_ingest.cli:main"
lexgraph-worker-segment = "services.svc_segment.cli:main"
lexgraph-worker-enrich = "services.svc_enrich.cli:main"
lexgraph-worker-graph = "services.svc_graph_write.cli:main"
lexgraph-worker-index = "services.svc_index.cli:main"
lexgraph-worker-retrieve = "services.svc_retrieve.cli:main"
lexgraph-worker-generate = "services.svc_generate.cli:main"
lexgraph-worker-eval = "services.svc_eval.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["services*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "N", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
